Move install files and change file to main.py
