+++ requirements.txt
@@ -41,7 +41,6 @@ six==1.16.0
 soupsieve==2.4
 sympy==1.11.1
 tokenizers==0.13.3
-#torch==2.4.1+cu118
 tqdm==4.65.0
 transformers==4.27.4
 triton==3.0.0
